@@ -4,30 +4,30 @@
   "author": "Markus Eriksson",
   "license": "ISC",
   "dependencies": {
-    "@types/request": "^2.47.0",
-    "@types/request-promise": "^4.1.41",
-    "@types/winston": "^2.3.9",
-    "mongodb": "^3.0.8",
-    "request": "^2.87.0",
-    "request-promise": "^4.2.2",
-    "rxjs": "^5.5.11",
-    "twit": "^2.2.9",
-    "winston": "^2.4.2"
+    "mongodb": "^3.1.8",
+    "request": "^2.88.0",
+    "request-promise-native": "^1.0.5",
+    "rxjs": "^6.3.3",
+    "twit": "^2.2.11",
+    "winston": "^3.1.0"
   },
   "devDependencies": {
-    "@types/aws-lambda": "^8.10.5",
+    "@types/request": "^2.47.1",
+    "@types/request-promise-native": "^1.0.15",
+    "@types/aws-lambda": "^8.10.14",
     "@types/jest": "^22.2.3",
-    "@types/mongodb": "^3.0.18",
-    "@types/node": "^9.6.18",
-    "@types/twit": "^2.2.12",
-    "jest": "^22.4.4",
-    "serverless-offline": "^3.24.3",
-    "serverless-webpack": "^5.1.5",
-    "ts-jest": "^22.4.6",
-    "ts-loader": "^4.3.0",
-    "tslint": "^5.10.0",
-    "typescript": "^2.8.3",
-    "webpack": "^4.9.2",
+    "@types/mongodb": "^3.1.12",
+    "@types/node": "^9.6.35",
+    "@types/twit": "^2.2.20",
+    "jest": "^23.6.0",
+    "serverless-offline": "^3.29.0",
+    "serverless-webpack": "^5.2.0",
+    "ts-jest": "^23.10.4",
+    "ts-loader": "^5.2.0",
+    "ts-node": "^7.0.1",
+    "tslint": "^5.11.0",
+    "typescript": "^3.1.3",
+    "webpack": "^4.23.1",
     "webpack-node-externals": "^1.7.2"
   },
   "scripts": {
