@@ -1,15 +1,16 @@
 import fs = require("fs");
-import rp = require("request-promise");
+import rp = require("request-promise-native");
 
 // Borrow Long from mongo as a 64 bit integer for tweet ids
 import { Long } from "mongodb";
-import { Logger, transports } from "winston";
+import { map } from 'rxjs/operators';
+import { createLogger, transports } from "winston";
 import { splitStream } from "./splitStream";
 import { St1Repository } from "./St1Repository";
 import { St1Tweet } from "./St1Tweet";
 import { ITweetDoc } from "./St1TwitterClient";
 
-const logger = new Logger({ transports: [ new transports.Console() ] });
+const logger = createLogger({ transports: [ new transports.Console() ] });
 
 const b95 = "b95";
 const diesel = "diesel";
@@ -62,16 +63,17 @@ export class St1 {
         const st1 = new St1();
 
         return new Promise((resolve, reject) => {
-            splitStream(readStream)
-                .map((row) => JSON.parse(row))
-                .subscribe(
-                    (st1TweetData) => st1.appendTweetFromData(st1TweetData),
-                    (err) => {
-                        logger.error("Failed to fill with data from stream, using cached data only. %s", err);
-                        resolve(st1);
-                    },
-                    () => resolve(st1),
-                );
+            splitStream(readStream).pipe(
+                map((row) => JSON.parse(row)),
+            )
+            .subscribe(
+                (st1TweetData) => st1.appendTweetFromData(st1TweetData),
+                (err) => {
+                    logger.error("Failed to fill with data from stream, using cached data only. %s", err);
+                    resolve(st1);
+                },
+                () => resolve(st1)
+            );
         });
     }
 
