@@ -1,4 +1,4 @@
-import { Observable } from "rxjs/Rx";
+import { Observable } from "rxjs";
 
 /**
  * Split the given stream on the given separator (default '\n').
