@@ -1,8 +1,8 @@
-import { Logger, transports  } from "winston";
+import { createLogger, transports  } from "winston";
 import { St1 } from "./St1";
 import { St1Repository } from "./St1Repository";
 
-const logger = new Logger({ transports: [ new transports.Console() ] });
+const logger = createLogger({ transports: [ new transports.Console() ] });
 
 const atlasUri: string = process.env.MONGODB_ATLAS_CLUSTER_URI_R || "";
 const st1Repo = new St1Repository(atlasUri);
